gstreamer: Remove plugins from previous GStreamer
We build a newer version of GStreamer in this runtime, and some plugins
have been renamed in between the two versions. We cannot have two plugins
that provide the same GType, since that will crash the plugin scanner. We
also cannot remove the old plugins, since they belong to the GNOME
runtime and are mounted readonly. This replaces them with empty files,
which will still cause the plugin scanner to blacklist them, but at least
the correct plugins will be blacklisted.

This could be more easily done with a source of "type": "shell", but the
manifest still needs to build with flatpak-builder 0.8 for the time
being.

https://phabricator.endlessm.com/T19103