Generate the SDK manifest using Python
We're doing two passes to generate the SDK manifest: first from the
Makefile and then from the generate-manifest Python script. The script,
incidentally, overwrites some of the substitutions we perform from the
Makefile.

Let's move the manifest munging to the Python script only.

We need to still use the Makefile pattern substitution for various
ancillary files, so we need to use a different extension for the
manifest, to avoid Make getting confused.