Add extension point for GTK+ modules
GTK+ modules provide extra functionality to GTK+, so this patch adds
the extension point to the freedesktop Platform and Sdk runtimes,
making it easy for those modules to be used with Flatpak apps.

Extensions that have GTK+ modules should have an org.gtk.Gtk3module as
a prefix in their name. The version for the extensions should the same
as GTK+'s (3.0).
Note that in order for the actual modules to be loaded, they have to
be added to the GTK3_MODULES environment variable (passed Flatpak when
running an app).

https://phabricator.endlessm.com/T23764