Add version information to runtime builds
This is passed to sdk/os-release.bst by way of a yml file which is
included by the element. Unfortunately, this implementation requires
that we add a file (flatpak-version.yml) which is tracked by git.

This might be tidier if buildstream allows us to add local variable
overrides, for example in a local project.conf.

https://phabricator.endlessm.com/T28284