Improve how we track git_tag sources copied from junctions
Instead of keeping the existing track and ref properties for git_tag
sources that are copied from gnome-sdk.bst and freedesktop-sdk.bst,
change "ref" properties to "track" with the equivalent commit hash. This
ensures that our version will always point to the same commit as the
upstream version, even after re-running `bst track`.

https://phabricator.endlessm.com/T28284