Specify junit reports directory for some projects
This is a bit of an ugly place to be doing this, but the junit reports
directory for our test suites is specified at configure time. With this
change, we won't be forced to create a special build to run unit tests,
and instead Jenkins can test with a regular build directory, potentially
with a pre-existing cached artifact. As a side-effect, test runners can
assume that test results are placed in %{build-root}/_reports.

https://phabricator.endlessm.com/T28284