Remove bundle data from appstream metainfo
On older flatpak as used in our infrastructure this causes a segfault
when `flatpak build-update-repo` tries to construct the global appstream
XML. As it turns out, flatpak tries to throw out any `<bundle>` nodes it
finds and always constructs a new one based on the flatpak metadata. So,
even though this information is correct, it's ignored.

https://phabricator.endlessm.com/T30093