project: Use upstream's artifact servers
We're not developing or building this project anymore, so we'd like to
get rid of the artifact server that this is the only user of. Instead,
just use upstream's artifact servers for GNOME SDK and Freedesktop SDK.
It will be probably be pointless since the software versions specified
here are likely long gone from those servers, but it's better than a
pile of errors trying to use a non-existent server.

https://phabricator.endlessm.com/T35728